+++ Assets/OxGKit/Utilities/Scripts/Runtime/Requester/Requester.cs
@@ -8,6 +8,13 @@ using UnityEngine.Networking;
 
 namespace OxGKit.Utilities.Requester
 {
+    public struct ErrorInfo
+    {
+        public string url;
+        public string message;
+        public Exception exception;
+    }
+
     public static class Requester
     {
         private static ARCCache<string, AudioClip> _arcAudios = null;
@@ -265,7 +272,7 @@ namespace OxGKit.Utilities.Requester
         /// <param name="cached"></param>
         /// <param name="timeoutSeconds"></param>
         /// <returns></returns>
-        public static async UniTask<AudioClip> RequestAudio(string url, AudioType audioType = AudioType.MPEG, Action<AudioClip> successAction = null, Action<string> errorAction = null, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
+        public static async UniTask<AudioClip> RequestAudio(string url, AudioType audioType = AudioType.MPEG, Action<AudioClip> successAction = null, Action<ErrorInfo> errorAction = null, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
         {
             if (string.IsNullOrEmpty(url))
             {
@@ -336,7 +343,7 @@ namespace OxGKit.Utilities.Requester
         /// <param name="cached"></param>
         /// <param name="timeoutSeconds"></param>
         /// <returns></returns>
-        public static async UniTask<Texture2D> RequestTexture2D(string url, Action<Texture2D> successAction = null, Action<string> errorAction = null, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
+        public static async UniTask<Texture2D> RequestTexture2D(string url, Action<Texture2D> successAction = null, Action<ErrorInfo> errorAction = null, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
         {
             if (string.IsNullOrEmpty(url))
             {
@@ -412,7 +419,7 @@ namespace OxGKit.Utilities.Requester
         /// <param name="cached"></param>
         /// <param name="timeoutSeconds"></param>
         /// <returns></returns>
-        public static async UniTask<Sprite> RequestSprite(string url, Action<Sprite> successAction = null, Action<string> errorAction = null, Vector2 position = default, Vector2 pivot = default, float pixelPerUnit = 100, uint extrude = 0, SpriteMeshType meshType = SpriteMeshType.FullRect, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
+        public static async UniTask<Sprite> RequestSprite(string url, Action<Sprite> successAction = null, Action<ErrorInfo> errorAction = null, Vector2 position = default, Vector2 pivot = default, float pixelPerUnit = 100, uint extrude = 0, SpriteMeshType meshType = SpriteMeshType.FullRect, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
         {
             var texture = await RequestTexture2D(url, null, errorAction, cts, cached, timeoutSeconds);
             if (texture != null)
@@ -434,7 +441,7 @@ namespace OxGKit.Utilities.Requester
         /// <param name="cts"></param>
         /// <param name="timeoutSeconds"></param>
         /// <returns></returns>
-        public static async UniTask<byte[]> RequestBytes(string url, Action<byte[]> successAction = null, Action<string> errorAction = null, CancellationTokenSource cts = null, int? timeoutSeconds = null)
+        public static async UniTask<byte[]> RequestBytes(string url, Action<byte[]> successAction = null, Action<ErrorInfo> errorAction = null, CancellationTokenSource cts = null, int? timeoutSeconds = null)
         {
             if (string.IsNullOrEmpty(url))
             {
@@ -466,7 +473,7 @@ namespace OxGKit.Utilities.Requester
         /// <param name="cached"></param>
         /// <param name="timeoutSeconds"></param>
         /// <returns></returns>
-        public static async UniTask<string> RequestText(string url, Action<string> successAction = null, Action<string> errorAction = null, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
+        public static async UniTask<string> RequestText(string url, Action<string> successAction = null, Action<ErrorInfo> errorAction = null, CancellationTokenSource cts = null, bool cached = true, int? timeoutSeconds = null)
         {
             if (string.IsNullOrEmpty(url))
             {
@@ -534,7 +541,7 @@ namespace OxGKit.Utilities.Requester
         /// <param name="processResponse"></param>
         /// <param name="errorAction"></param>
         /// <returns></returns>
-        internal static async UniTask<T> SendRequest<T>(string url, CancellationTokenSource cts, int? timeoutSeconds, Func<UnityWebRequest> createRequest, Func<UnityWebRequest, T> processResponse, Action<string> errorAction)
+        internal static async UniTask<T> SendRequest<T>(string url, CancellationTokenSource cts, int? timeoutSeconds, Func<UnityWebRequest> createRequest, Func<UnityWebRequest, T> processResponse, Action<ErrorInfo> errorAction)
         {
             UnityWebRequest request = null;
             try
@@ -557,10 +564,13 @@ namespace OxGKit.Utilities.Requester
                     request.result == UnityWebRequest.Result.ProtocolError ||
                     request.result == UnityWebRequest.Result.ConnectionError)
                 {
-                    string errorMsg = request.error;
-                    errorAction?.Invoke(errorMsg);
+                    var errorInfo = new ErrorInfo();
+                    errorInfo.url = url;
+                    errorInfo.message = request.error;
+                    errorInfo.exception = null;
+                    errorAction?.Invoke(errorInfo);
                     request.Dispose();
-                    Logging.PrintWarning<Logger>($"<color=#FF0000>Request failed. URL: {url}, ErrorMsg: {errorMsg}</color>");
+                    Logging.PrintWarning<Logger>($"<color=#FF0000>Request failed. URL: {errorInfo.url}, ErrorMsg: {errorInfo.message}</color>");
                     return default;
                 }
 
@@ -576,10 +586,13 @@ namespace OxGKit.Utilities.Requester
             }
             catch (Exception ex)
             {
-                string errorMsg = string.IsNullOrEmpty(request?.error) ? $"Request failed. URL: {url}, Exception: {ex}" : request.error;
-                errorAction?.Invoke(errorMsg);
+                var errorInfo = new ErrorInfo();
+                errorInfo.url = url;
+                errorInfo.message = request?.error;
+                errorInfo.exception = ex;
+                errorAction?.Invoke(errorInfo);
                 request?.Dispose();
-                Logging.PrintWarning<Logger>($"<color=#FF0000>{errorMsg}</color>");
+                Logging.PrintWarning<Logger>($"<color=#FF0000>Request failed. URL: {errorInfo.url}, ErrorMsg: {errorInfo.message}, Exception: {ex}</color>");
                 return default;
             }
         }