+++ Assets/OxGKit/Utilities/Scripts/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## CHANGELOG
 
 ## [1.4.3] - 2025-03-25
-- Modified Requester's errorAction to return error messages.
+- Modified Requester's errorAction to return ErrorInfo.
 
 ## [1.4.2] - 2025-03-21
 - Fixed CancellationTokenSource disposal bug.